feat: cast to storage type

+++ cv/common/nn.py
@@ -31,8 +31,8 @@ class BatchNorm:
     batch_mean, batch_var = self.calc_stats(xd)
     # NOTE: wow, this is done all throughout training in most PyTorch models
     if self.track_running_stats and Tensor.training:
-      self.running_mean.assign((1-self.momentum) * self.running_mean + self.momentum * batch_mean.detach().cast(dtypes.float32))
-      self.running_var.assign((1-self.momentum) * self.running_var + self.momentum * xd.numel()/(xd.numel()-xd.shape[1]) * batch_var.detach().cast(dtypes.float32))
+      self.running_mean.assign((1-self.momentum) * self.running_mean + self.momentum * batch_mean.detach().cast(self.running_mean.dtype))
+      self.running_var.assign((1-self.momentum) * self.running_var + self.momentum * xd.numel()/(xd.numel()-xd.shape[1]) * batch_var.detach().cast(self.running_var.dtype))
       self.num_batches_tracked += 1
     return x.batchnorm(self.weight, self.bias, batch_mean, batch_var.add(self.eps).rsqrt()).cast(x.dtype)
